[API][Account] Adjust password change to chosen structure
