+++ src/Sylius/Behat/Resources/config/suites.yml
@@ -5,7 +5,6 @@ imports:
     - suites/api/account/customer.yml
     - suites/api/account/customer_registration.yml
     - suites/api/account/login.yml
-    - suites/api/account/shop_user_account.yml
     - suites/api/addressing/managing_countries.yml
     - suites/api/addressing/managing_zones.yml
     - suites/api/account/address_book.yml
@@ -52,7 +51,6 @@ imports:
     - suites/ui/account/customer_registration.yml
     - suites/ui/account/email_verification.yml
     - suites/ui/account/login.yml
-    - suites/ui/account/shop_user_account.yml
     - suites/ui/addressing/managing_countries.yml
     - suites/ui/addressing/managing_zones.yml
     - suites/ui/admin/dashboard.yml
