--- src/Sylius/Behat/Resources/config/suites/api/account/shop_user_account.yml
@@ -1,17 +1,0 @@
-# This file is part of the Sylius package.
-# (c) Paweł Jędrzejewski
-
-default:
-    suites:
-        api_shop_user_account:
-            contexts:
-                - sylius.behat.context.hook.doctrine_orm
-
-                - sylius.behat.context.setup.channel
-                - sylius.behat.context.setup.customer
-                - sylius.behat.context.setup.shop_api_security
-
-                - sylius.behat.context.api.shop.account
-
-            filters:
-                tags: "@shop_user_account && @api"
