+++ src/Sylius/Bundle/ApiBundle/Resources/config/serialization/Commands/ChangePasswordShopUser.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" ?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<serializer xmlns="http://symfony.com/schema/dic/serializer-mapping"
+            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
+>
+    <class name="Sylius\Bundle\ApiBundle\Command\ChangePasswordShopUser">
+        <attribute name="oldPassword">
+            <group>customer:password:write</group>
+        </attribute>
+        <attribute name="password">
+            <group>customer:password:write</group>
+        </attribute>
+        <attribute name="confirmPassword">
+            <group>customer:password:write</group>
+        </attribute>
+    </class>
+</serializer>
